make proportion_from_point work
