+++ manim/mobject/shape_matchers.py
@@ -4,6 +4,8 @@ __all__ = ["SurroundingRectangle", "BackgroundRectangle", "Cross", "Underline"]
 
 from typing import Optional
 
+from manim.utils.color import Color
+
 from ..constants import *
 from ..mobject.geometry import Line, RoundedRectangle
 from ..mobject.mobject import Mobject
